Remove flag `dashboard-redesign` and purge old code
