--- frontend/src/lib/hooks/usePrevious.js
@@ -1,9 +1,0 @@
-import { useRef, useEffect } from 'react'
-
-export function usePrevious(value) {
-    const ref = useRef()
-    useEffect(() => {
-        ref.current = value
-    })
-    return ref.current
-}
