+++ frontend/src/scenes/dashboard/Dashboard.tsx
@@ -2,7 +2,6 @@ import React, { useEffect } from 'react'
 import { SceneLoading } from 'lib/utils'
 import { BindLogic, useActions, useValues } from 'kea'
 import { dashboardLogic, DashboardLogicProps } from 'scenes/dashboard/dashboardLogic'
-import { DashboardHeader } from 'scenes/dashboard/DashboardHeader'
 import { DashboardItems } from 'scenes/dashboard/DashboardItems'
 import { dashboardsModel } from '~/models/dashboardsModel'
 import { DateFilter } from 'lib/components/DateFilter/DateFilter'
@@ -17,9 +16,7 @@ import { NotFound } from 'lib/components/NotFound'
 import { DashboardReloadAction, LastRefreshText } from 'scenes/dashboard/DashboardReloadAction'
 import { SceneExport } from 'scenes/sceneTypes'
 import { InsightErrorState } from 'scenes/insights/EmptyStates'
-import { LemonDashboardHeader } from './LemonDashboardHeader'
-import { FEATURE_FLAGS } from 'lib/constants'
-import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
+import { DashboardHeader } from './DashboardHeader'
 
 interface Props {
     id?: string
@@ -53,7 +50,6 @@ function DashboardView(): JSX.Element {
     } = useValues(dashboardLogic)
     const { dashboardsLoading } = useValues(dashboardsModel)
     const { setDashboardMode, setDates, reportDashboardViewed } = useActions(dashboardLogic)
-    const { featureFlags } = useValues(featureFlagLogic)
 
     useEffect(() => {
         reportDashboardViewed()
@@ -98,9 +94,7 @@ function DashboardView(): JSX.Element {
 
     return (
         <div className="dashboard">
-            {dashboardMode !== DashboardMode.Public &&
-                dashboardMode !== DashboardMode.Internal &&
-                (featureFlags[FEATURE_FLAGS.DASHBOARD_REDESIGN] ? <LemonDashboardHeader /> : <DashboardHeader />)}
+            {dashboardMode !== DashboardMode.Public && dashboardMode !== DashboardMode.Internal && <DashboardHeader />}
             {receivedErrorsFromAPI ? (
                 <InsightErrorState title="There was an error loading this dashboard" />
             ) : !items || items.length === 0 ? (
