+++ frontend/src/scenes/funnels/FunnelBarGraph.scss
@@ -353,7 +353,7 @@ $funnel_canvas_background: #fff;
 
 // Overrides for dashboard styles
 
-.dashboard-item-content {
+.InsightCard {
     .funnel-bar-graph {
         margin: 0;
         padding: 12px 0;
