+++ frontend/src/scenes/funnels/FunnelHistogram.scss
@@ -1,4 +1,4 @@
-.dashboard-item-content {
+.InsightCard {
     .histogram-container {
         max-height: 100%;
     }
