+++ frontend/src/scenes/insights/InsightTabs/FunnelTab/FunnelStepTable.scss
@@ -1,6 +1,5 @@
 @import '~/vars';
 
-.dashboard-item-container,
 .InsightCard {
     .funnel-steps-table {
         margin: 0 !important;
