--- frontend/src/scenes/insights/SaveModal/index.ts
@@ -1,1 +1,0 @@
-export * from './SaveModal'
