+++ frontend/src/scenes/saved-insights/SavedInsights.scss
@@ -38,42 +38,6 @@
         align-items: center;
     }
 
-    .dashboard-item {
-        padding-bottom: 10px;
-        cursor: pointer;
-
-        .dashboard-item-container:not(.funnel) .dashboard-item-content {
-            transform: scale(0.333);
-            transform-origin: top left;
-            overflow: inherit;
-            position: inherit;
-            & > div {
-                height: 300% !important;
-                width: 300% !important;
-            }
-        }
-        .dashboard-item-header {
-            cursor: pointer !important;
-        }
-
-        .svg-funnel-js__label * {
-            font-size: 8px;
-        }
-        &:hover {
-            border-color: var(--primary);
-        }
-    }
-
-    .dashboard-item-footer {
-        bottom: 0;
-        position: absolute;
-        background: #fff;
-        padding: 0.25rem;
-        border-top: 1px solid var(--muted);
-        width: 100%;
-        font-size: 11px;
-    }
-
     .saved-insight-empty-state {
         display: flex;
         justify-content: center;
