+++ frontend/src/scenes/trends/viz/ActionsPie.scss
@@ -36,96 +36,3 @@
         }
     }
 }
-
-.dashboard-item.di-width-1,
-.dashboard-item.di-width-2 {
-    .actions-pie-component {
-        .pie-chart {
-            height: calc(100% - 20px);
-        }
-
-        h1 {
-            font-size: 16px;
-            .label {
-                display: none;
-            }
-        }
-    }
-}
-
-.dashboard-item.di-width-3,
-.dashboard-item.di-width-4 {
-    .actions-pie-component {
-        .pie-chart {
-            height: calc(100% - 30px);
-        }
-
-        h1 {
-            font-size: 24px;
-            .label {
-                display: none;
-            }
-        }
-    }
-}
-
-.dashboard-item.di-width-5,
-.dashboard-item.di-width-6 {
-    .actions-pie-component {
-        .pie-chart {
-            height: calc(100% - 40px);
-        }
-
-        h1 {
-            font-size: 36px;
-        }
-    }
-}
-.dashboard-item.di-width-7,
-.dashboard-item.di-width-8,
-.dashboard-item.di-width-9,
-.dashboard-item.di-width-10 {
-    .actions-pie-component {
-        .pie-chart {
-            height: calc(100% - 70px);
-        }
-
-        h1 {
-            font-size: 48px;
-        }
-    }
-}
-.dashboard-item.di-width-11,
-.dashboard-item.di-width-12,
-.dashboard-item.di-width-13,
-.dashboard-item.di-width-14,
-.dashboard-item.di-width-15,
-.dashboard-item.di-width-16 {
-    .actions-pie-component {
-        .pie-chart {
-            height: calc(100% - 100px);
-        }
-
-        h1 {
-            font-size: 64px;
-        }
-    }
-}
-.dashboard-item.di-width-17,
-.dashboard-item.di-width-18,
-.dashboard-item.di-width-19,
-.dashboard-item.di-width-20,
-.dashboard-item.di-width-21,
-.dashboard-item.di-width-22,
-.dashboard-item.di-width-23,
-.dashboard-item.di-width-24 {
-    .actions-pie-component {
-        .pie-chart {
-            height: calc(100% - 130px);
-        }
-
-        h1 {
-            font-size: 96px;
-        }
-    }
-}
